requirements.txt
aenum~=3.1.11
aiomysql~=0.1.1
aioredis~=2.0.1
AlbertUnruhUtils.py~=2.3.0
NAFF[voice,speedup]~=1.12.0
python-datetimeparser==0.4.rc  # fix version since 0.5(rc) might be completely breaking
python-dotenv~=0.21.0
PyYAML~=6.0
SQLAlchemy~=1.4.42
